:sparkles: feat: criteria & app tabs
